// KInvFront/src/app/ServiceSwagger/model/empleadoDto.ts
/**
 * Backend
 * Descripción
 *
 * OpenAPI spec version: 0.0.1
 * Contact: willypanza11@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface EmpleadoDto { 
    estado?: string;
    estadoVacunacion?: string;
    idEmpleado?: number;
    salario?: number;
    tipo?: string;
}